+++ src/js/genghis/views/keyboard_shortcuts.js
@@ -6,24 +6,24 @@ Genghis.Views.KeyboardShortcuts = Backbone.View.extend({
     },
     initialize: function() {
         _.bindAll(this, 'render', 'show', 'hide', 'toggle');
-        $(document).bind('keyup', 'shift+/', this.toggle);
+        Mousetrap.bind('?', this.toggle);
         $('footer a.keyboard-shortcuts').click(this.show);
         this.render();
     },
     render: function() {
-        $(this.el).html(this.template.render()).modal({backdrop: true, keyboard: true, show: false});
+        this.$el.html(this.template.render()).modal({backdrop: true, keyboard: true, show: false});
 
         return this;
     },
     show: function(e) {
         e.preventDefault();
-        $(this.el).modal('show');
+        this.$el.modal('show');
     },
     hide: function(e) {
         e.preventDefault();
-        $(this.el).modal('hide');
+        this.$el.modal('hide');
     },
     toggle: function() {
-        $(this.el).modal('toggle');
+        this.$el.modal('toggle');
     }
 });